# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: JudeOS — принятый онлайн-контракт и каркас
  version: 0.1.0
  description: 'Контракт #16 принят пользователем (DECISIONS U2026-10-07-S0-01, PR #78 merged). Семь бизнес-операций запланированы; #19 реализует только служебные операции каркаса. Все данные синтетические. Источник контрактов — этот файл; runtime /openapi.json содержит только x-status implemented. Правила — api/README.md.'
  contact:
    url: https://github.com/StarMadeGalaxy/JudeOS/issues/16
servers:
- url: https://judeos.example
  description: Зарезервированный пример, не работающий сервис.
...
security:
- staffSession: []
paths:
...
  /openapi.json:
    get:
      operationId: getRuntimeOpenAPI
      tags:
      - operations
      summary: Получить контракт работающего каркаса
      description: Bundled OpenAPI 3.0.3 только реализованных операций. Семь запланированных бизнес-операций
        отсутствуют в выдаче; полный принятый контракт в репозитории. Публичный dev/synthetic маршрут, без
        CSRF/tenant/мутаций.
      x-status: implemented
      x-issue: 19
      security: []
      responses:
        '200':
          description: Получить контракт работающего каркаса
          headers: *scaffoldid001
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
              example:
                openapi: 3.0.3
                info:
                  title: JudeOS scaffold
                  version: 0.1.0
                paths: {}
        '404':
          description: Путь не найден
          headers: *scaffoldid001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_FOUND
                message: Путь не найден
                request_id: '00000000000000000000000000000001'
  /docs:
    get:
      operationId: getSwaggerUI
      tags:
      - operations
      summary: Открыть Swagger UI каркаса
      description: 'Локальные CSS/JS и /openapi.json, Try it out отключён. Только dev/synthetic; production-доступ
        — #22. Публичный маршрут, без CSRF/tenant/мутаций. /docs/ с завершающим слешем — 404, редиректов
        нет.'
      x-status: implemented
      x-issue: 19
      security: []
      responses:
        '200':
          description: Открыть Swagger UI каркаса
          headers: *scaffoldid001
          content:
            text/html:
              schema:
                type: string
              example: <!doctype html><html lang="ru"><title>JudeOS API</title><body>Swagger UI</body></html>
        '404':
          description: Путь не найден
          headers: *scaffoldid001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_FOUND
                message: Путь не найден
                request_id: '00000000000000000000000000000001'
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: __Host-judeos-session
      description: Серверная отзываемая сессия. Secure/HttpOnly/SameSite=Lax/Path=/;
        host-only; секрет не передаётся в JSON и не хранится в localStorage/IndexedDB.
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          enum:
          - INVALID_REQUEST
          - UNAUTHENTICATED
          - LOGIN_FAILED
          - CSRF_REJECTED
          - FORBIDDEN
          - NOT_FOUND
          - METHOD_NOT_ALLOWED
          - UNSUPPORTED_MEDIA_TYPE
          - PAYLOAD_TOO_LARGE
          - OPERATION_ID_REUSED
          - SESSION_CANCELLED
          - ATHLETE_ARCHIVED
          - RATE_LIMITED
          - INTERNAL_ERROR
          - SERVICE_UNAVAILABLE
        message:
          type: string
          minLength: 1
          maxLength: 300
        request_id:
          type: string
          minLength: 1
          maxLength: 100
...